+++ database/nostr-lmdb/src/store/error.rs
@@ -22,12 +22,14 @@ pub enum Error {
     Key(key::Error),
     Secp256k1(secp256k1::Error),
     OneshotRecv(oneshot::error::RecvError),
-    /// MPSC send error
-    MpscSend,
+    /// Flume channel send error
+    FlumeSend,
     /// The event kind is wrong
     WrongEventKind,
     /// Not found
     NotFound,
+    /// Batched transaction failed - sent to operations that didn't cause the error
+    BatchTransactionFailed,
 }
 
 impl std::error::Error for Error {}
@@ -42,9 +44,10 @@ impl fmt::Display for Error {
             Self::Key(e) => write!(f, "{e}"),
             Self::Secp256k1(e) => write!(f, "{e}"),
             Self::OneshotRecv(e) => write!(f, "{e}"),
-            Self::MpscSend => write!(f, "mpsc channel send error"),
+            Self::FlumeSend => write!(f, "flume channel send error"),
             Self::NotFound => write!(f, "Not found"),
             Self::WrongEventKind => write!(f, "Wrong event kind"),
+            Self::BatchTransactionFailed => write!(f, "Batched transaction failed"),
         }
     }
 }